feat: Implement ConfigManager (Task 3)
Introduce ConfigManager to centralize configuration loading and validation. Refactor cli.ts to use ConfigManager.
